// src/interfaces/comic-metadata.interface.ts
export interface ComicMetadata {
    id: number | null; // This will be set after inserting the metadata
    title: string | null; // Title can be null if not set
    series: string;
    number: string | null; // Number can be null if not set
    count: number | null;
    volume: string | null;
    alternateSeries: string | null;
    alternateNumber: string | null;
    alternateVolume: string | null;
    summary: string | null;
    notes: string | null;
    year: number | null;
    month: number | null;
    day: number | null;
    website: string | null;
    page_count: number | null;
    language: string | null;
    format: string | null;
    blackAndWhite: boolean | null;
    manga: boolean | null;
    mangaDirection: string | null; // Added for manga direction
    scanInfo: string | null;
    storyArcId: number | null;
    communityRating: number | null;
    mainCharacter : string | null;
    review: string | null;
    gtin: string | null;
}
// ...
export type ComicMetadataInsert = Omit<ComicMetadata, 'id'> & { id?: number | null };